svg/painting/parsing/stroke-dasharray-invalid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing stroke-dasharray with invalid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeDasharrayProperty" rel="help"/><h:meta content="stroke-dasharray supports only the grammar 'none | dasharray'" name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_invalid_value("stroke-dasharray", "auto");
test_invalid_value("stroke-dasharray", "none 10px");
test_invalid_value("stroke-dasharray", "20px / 30px");
test_invalid_value("stroke-dasharray", "-40px");

  ]]></script></svg>
